fix : ID Long type 변경

+++ src/main/java/knu/chcse/knucseofficialserver/domain/entity/attachment/AttachmentJpaRepository.java
@@ -2,5 +2,5 @@ package knu.chcse.knucseofficialserver.domain.entity.attachment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AttachmentJpaRepository extends JpaRepository<Attachment, Integer> {
+public interface AttachmentJpaRepository extends JpaRepository<Attachment, Long> {
 }
